Update lookup table generation, fix some minor typing issues in float.

+++ Sources/SF2Lib/include/SF2Lib/Entity/Generator/Definition.hpp
@@ -5,7 +5,7 @@
 #include <array>
 #include <string>
 
-#include "SF2Lib/DSP/DSP.hpp"
+#include "SF2Lib/DSP.hpp"
 #include "SF2Lib/Entity/Generator/Amount.hpp"
 #include "SF2Lib/Entity/Generator/Index.hpp"
 
@@ -137,11 +137,11 @@ public:
   Float convertedValueOf(const Amount& amount) const noexcept {
     switch (valueKind_) {
       case ValueKind::coarseOffset: return valueOf(amount) * 32768;
-      case ValueKind::signedCents: return valueOf(amount) / 1200.0;
+      case ValueKind::signedCents: return Float(valueOf(amount) / 1200.0f);
 
       case ValueKind::signedCentsBel:
       case ValueKind::unsignedPercent:
-      case ValueKind::signedPercent: return valueOf(amount) / 10.0;
+      case ValueKind::signedPercent: return valueOf(amount) / 10.0f;
 
       case ValueKind::signedFrequencyCents: return Float(DSP::centsToFrequency(valueOf(amount)));
       case ValueKind::signedTimeCents: return DSP::centsToSeconds(valueOf(amount));
